requirements.txt
gdax==1.0.6
python-binance==0.5.12
